Update version numbers and test counts for multiple crates to 0.3.6 (2026-07-08)
- oxiarc-lz4: Incremented version to 0.3.6, updated test count to 140, added new tests for dict_block_roundtrip and proptest_roundtrip.
- oxiarc-lzhuf: Incremented version to 0.3.6, updated test count to 175, added fuzzing tests and improved edge case handling.
- oxiarc-lzma: Incremented version to 0.3.6, updated test count to 156, fixed multi-chunk encoder/decoder desync and improved error handling.
- oxiarc-lzw: Incremented version to 0.3.6, updated test count to 79, added property-based round-trip testing and default implementation for LzwConfig.
- oxiarc-snappy: Incremented version to 0.3.6, updated test count to 114, added new proptest-based round-trip regression suite and example for pooled streaming.
- oxiarc-szip: Incremented version to 0.3.6, updated test count to 27, added default implementation for SzipParams and improved error handling.
- oxiarc-zstd: Incremented version to 0.3.6, updated test count to 186, hardened frame decoding and marked BlockType/LiteralsBlockType as #[non_exhaustive].

+++ fuzz/fuzz_targets/fuzz_szip_decode.rs
@@ -40,7 +40,7 @@ fuzz_target!(|data: &[u8]| {
     // Cap sample count so a bogus header can't force an unbounded
     // allocation attempt in the decoder's output buffer.
     let samples = (samples_raw as usize) % 65_536;
-    let reference_sample_interval = u32::from(rsi_tag) * u32::from(pixels_per_block);
+    let reference_sample_interval = u32::from(rsi_tag) * pixels_per_block;
 
     let params = SzipParams {
         bits_per_pixel,
